HW2/assignment/requirements.txt
Cython>=0.27.3
Jinja2>=2.7.3
MarkupSafe>=0.23
Pillow>=2.7.0
backports.ssl-match-hostname>=3.4.0.2
certifi>=14.05.14
gnureadline>=6.3.3
ipython>=3.2.1
matplotlib>=1.4.2
mock>=1.0.1
nose>=1.3.4
numpy>=1.9.1
pyparsing>=2.0.3
python-dateutil>=2.4.0
pytz>=2014.10
pyzmq>=14.4.1
scipy==1.1.0
six>=1.9.0
tornado>=4.0.2
jupyterlab==2.2.6
torch
torchvision
